requirements.txt
flask
python-dotenv
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
binance-common==3.4.0
binance-sdk-spot==6.1.0
black==25.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
frozenlist==1.8.0
idna==3.11
imbalanced-learn==0.14.1
joblib==1.5.3
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
platformdirs==4.5.1
propcache==0.4.1
psycopg==3.3.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
requests==2.32.5
ruff==0.12.12
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sklearn-compat==0.1.5
SQLAlchemy==2.0.45
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
